src/top10_news_crawler/pyproject.toml
[project]
name = "top10-news-crawler"
version = "0.1.0"
description = "Crawl today’s top headlines from major international news sites, deduplicate and rank them, and generate a personalized Top 10 news digest."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.32.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=5.2.0",
  "PyYAML>=6.0.1",
  "rapidfuzz>=3.9.0",
  "python-dotenv>=1.0.1",
  "tenacity>=8.3.0"
]

[project.scripts]
top10 = "top10_news_crawler.main:cli"

[tool.ruff]
line-length = 100
